LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-focus-previous-outside-expected.txt

button 1 button 2

FAIL Focus should not be restored if the currently focused element is not inside the dialog. assert_equals: expected Element node <button id="b2">button 2</button> but got Element node <button id="b1">button 1</button>
FAIL Focus restore should not occur when the focused element is in a shadowroot outside of the dialog. assert_equals: document.activeElement should point at the shadow host. expected Element node <div id="host">

</div> but got Element node <button id="b2">button 2</button>
PASS Focus restore should occur when the focused element is in a shadowroot inside the dialog.
PASS Focus restore should occur when the focused element is slotted into a dialog.
PASS Focus restore should always occur for modal dialogs.
